selesai fitur user dan todolist_simple tinggal di test 	modified:   app.mjs 	modified:   controllers/todolist_simple.controllers.mjs 	modified:   controllers/user.controllers.mjs 	new file:   middlewares/error.handler.mjs 	deleted:    middlewares/error.handling.middleware.mjs 	modified:   middlewares/todolist_simple.middlewares.mjs 	modified:   middlewares/user.middlewares.mjs 	modified:   repositories/todolist_simple/delete.many.mjs 	modified:   repositories/todolist_simple/delete.one.mjs 	modified:   repositories/todolist_simple/find.by.id.mjs 	modified:   repositories/todolist_simple/index.mjs 	modified:   repositories/todolist_simple/insert.one.mjs 	modified:   repositories/todolist_simple/update.full.by.id.mjs 	modified:   repositories/todolist_simple/update.one.is_done.by.id.mjs 	modified:   repositories/todolist_simple/update.title.description.by.id.mjs 	modified:   routes/todolist_simple.routes.mjs 	modified:   routes/user.routes.mjs 	modified:   services/todolist_simple.services.mjs 	modified:   services/user.services.mjs 	modified:   utils/generateLoginToken.mjs 	modified:   utils/isNullish.mjs 	modified:   utils/validation.helper.mjs

+++ repositories/todolist_simple/delete.many.mjs
@@ -12,11 +12,11 @@ function sql(ids = []) {
   if (placeholders.length === 0)
     return `DELETE FROM todolist_simple WHERE id IN (NULL)`;
 
-  const sql = `DELETE FROM todolist_simple WHERE id IN (${placeholders});`;
+  const sql = `DELETE FROM todolist_simple WHERE user_id = ? AND id IN (${placeholders});`;
   return sql;
 }
 
-export default async function deleteMany({ ids = [] }) {
+export default async function deleteMany({ user_id, ids = [] }) {
   // Tetap menggunakan fungsi sql() yang sudah direvisi
-  return await pool.execute(sql(ids), ids);
+  return await pool.execute(sql(ids), [user_id, ...ids]);
 }
