+++ repositories/todolist_simple/delete.one.mjs
@@ -1,7 +1,7 @@
 import { pool } from "../../configs/database.connection.config.mjs";
 
-const sql = `DELETE FROM todolist_simple WHERE id = ? ;`;
+const sql = `DELETE FROM todolist_simple WHERE user_id = ? AND id = ? ;`;
 
-export default async function deleteOne({ id }) {
-  return await pool.execute(sql, [id]);
+export default async function deleteOne({ user_id, id }) {
+  return await pool.execute(sql, [user_id, id]);
 }
